Transforma projeto em Typescript
