+++ package.json
@@ -2,6 +2,7 @@
   "name": "medicos-crm-api",
   "version": "0.0.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "nodemon",
     "start": "node ./server.js"
@@ -11,7 +12,8 @@
     "debug": "~2.6.9",
     "dotenv-safe": "^8.2.0",
     "express": "~4.16.1",
-    "morgan": "~1.9.1"
+    "morgan": "~1.9.1",
+    "ts-node": "^10.4.0"
   },
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "^5.3.1",
