+++ routes/medicos.ts
@@ -1,6 +1,4 @@
-import { Router } from 'express';
-
-const router = Router();
+const router = require('express').Router();
 
 router.get('/:crm', (req, res) => res.send({ message: 'Hello nurse!' }));
